+++ pygnparser/pygnparser/gnparser.py
@@ -1,11 +1,12 @@
 from pygnparser.utils.pygnparser_utils import gnp_post, gnp_url
 
 
-def gnparser(names, with_details='on', cultivars='off', diaereses='off'):
+def gnparser(names, code='', with_details='on', cultivars='off', diaereses='off'):
     """
     Parse scientific names
 
     :param names: [str] Human name(s) separated by \\r\\n
+    :param code: [str] Specify the nomenclatural code (bacterial, botanical, cultivar, or zoological)
     :param with_details: [str] Turn detailed information about the name on or off
     :param cultivars: [str] Turn cultivar handling on or off
     :param diaereses: [str] Turn diaereses handling on or off
@@ -18,6 +19,7 @@ def gnparser(names, with_details='on', cultivars='off', diaereses='off'):
     url = gnp_url()
     args = {
         "names": names,
+        "code": code,
         "with_details": with_details,
         "cultivars": cultivars,
         "diaereses": diaereses,